requirements.txt
oandapyV20
pandas
numpy
scikit-learn
python-dotenv
requests
feedparser
google-generativeai
